Swapped CSRF initializer for a Mixin
For projects that don't use jQuery, the initializer trying to set `$.ajaxPrefilter` is a show-stopper. Instead of having this as part of an initializer, I opted to follow `ember-simple-auth`'s lead in creating a mixin for an Ember Data adapter. While adding the extra step of including the mixin to an application's adapter, this approach would additionally allow for an app to conditionally add the CSRF token when making requests of servers other than its Rails API.

(cherry picked from commit b5bdafa6e12ccd1480db8f3f518a14abf98d2243)